+++ src/components/shared/CardGrid/CardGrid.tsx
@@ -1,14 +1,16 @@
 import * as React from "react";
-import { Box, Grid, LayoutTypes } from "@nypl/design-system-react-components";
-import Image from "./../../shared/Image";
-import { ImageType } from "../Image/ImageTypes";
+import { Box, Grid } from "@nypl/design-system-react-components";
 import CardGridHeader from "./CardGridHeader";
 import Card from "./Card";
+import Image from "./../../shared/Image";
+import { ImageType } from "../Image/ImageTypes";
+
+export type CardGridLayoutTypes = "row" | "column" | "column_two_featured";
 
 export interface CardGridCommonProps {
   id: string;
   type: string;
-  title: string;
+  title?: string;
   headingColor?: string;
   description?: string;
   link?: string;
@@ -20,7 +22,7 @@ export interface CardGridCommonProps {
 type CardGridConditionalProps =
   | {
       items?: CardItem[];
-      layout?: LayoutTypes;
+      layout?: CardGridLayoutTypes;
       children?: never;
     }
   | {
@@ -47,23 +49,64 @@ export default function CardGrid({
   description,
   link,
   hrefText,
-  layout,
+  layout = "column",
   isBordered = false,
   isCentered = false,
   items,
   children,
 }: CardGridProps) {
-  let templateColumns = "repeat(auto-fit, minmax(300px, 1fr))";
+  // Logic for template columns based on layout.
+  let templateColumns = "repeat(12, 1fr)";
+  // Logic for template row based on layout.
   let isBorderedFinal = isBordered;
   let isCenteredFinal = isCentered;
   if (layout === "row") {
     templateColumns = "repeat(1, 1fr)";
-    isBorderedFinal = true;
     isCenteredFinal = true;
   }
 
+  // Helper function to get the grid column value based on layout and index.
+  function getGridColumn(
+    itemsCount: number,
+    layout: CardGridLayoutTypes,
+    index?: number
+  ) {
+    // Logic for grid column of each grid item.
+    // Default for 3 cards across, 4 + 4 + 4 = 12.
+    let gridColumn = "auto / span 4";
+    // Default logic for items count and adjusting the grid-column value based on number of cards.
+    // 5 cards.
+    if (itemsCount === 5 && index !== undefined) {
+      // 5 cards, 2 3
+      gridColumn = index < 2 ? "auto / span 6" : "auto / span 4";
+    }
+    // 4 cards.
+    if (itemsCount === 4) {
+      // 4 cards, 4
+      gridColumn = "auto / span 3";
+    }
+    // 2 cards.
+    if (itemsCount === 2) {
+      // 2 cards, 2
+      gridColumn = "auto / span 6";
+    }
+    // Column: two featured.
+    if (layout === "column_two_featured" && index !== undefined) {
+      if (itemsCount === 6) {
+        // 6 cards, 2 4
+        gridColumn = index < 2 ? "auto / span 6" : "auto / span 3";
+      }
+      if (itemsCount === 4) {
+        // 4 cards, 2 2
+        gridColumn = "auto / span 6";
+      }
+    }
+
+    return gridColumn;
+  }
+
   return (
-    <Box id={`${type}-${id}`} mb="2em">
+    <Box id={`${type}-${id}`} mb="xl">
       {title && (
         <CardGridHeader
           id={id}
@@ -74,47 +117,51 @@ export default function CardGrid({
         />
       )}
       {description && <p>{description}</p>}
-      {items && (
-        <Grid
-          as="ul"
-          listStyleType="none"
-          templateColumns={templateColumns}
-          gap="m"
-        >
-          {items.map((item: CardItem) => (
-            <li key={item.id}>
-              <Card
-                id={item.id}
-                heading={item.title}
-                description={item.description}
-                href={item.link}
-                layout={layout}
-                isBordered={isBorderedFinal}
-                isCentered={isCenteredFinal}
-                {...(item.image && {
-                  image: (
-                    <Image
-                      id={item.image.id}
-                      alt={item.image.alt}
-                      uri={item.image.uri}
-                      useTransformation={true}
-                      transformations={item.image.transformations}
-                      transformationLabel={"2_1_960"}
-                      layout="responsive"
-                      // width={item.image.width}
-                      // height={item.image.height}
-                      width={900}
-                      height={450}
-                      quality={90}
-                    />
-                  ),
-                })}
-              />
-            </li>
-          ))}
-        </Grid>
-      )}
-      {!items && children && <div>{children}</div>}
+      <Grid
+        as="ul"
+        listStyleType="none"
+        templateColumns={{ lg: templateColumns }}
+        gap="m"
+      >
+        {items &&
+          items.map((item: CardItem, index) => {
+            const { id, title, description, link, image } = item;
+            return (
+              <Box
+                as="li"
+                key={id}
+                gridColumn={getGridColumn(items.length, layout, index)}
+              >
+                <Card
+                  id={id}
+                  heading={title}
+                  description={description}
+                  href={link}
+                  layout={layout === "row" ? "row" : "column"}
+                  isBordered={isBorderedFinal}
+                  isCentered={isCenteredFinal}
+                  {...(image && {
+                    image: (
+                      <Image
+                        id={image.id}
+                        alt={image.alt}
+                        uri={image.uri}
+                        useTransformation={true}
+                        transformations={image.transformations}
+                        transformationLabel={"2_1_960"}
+                        layout="responsive"
+                        width={900}
+                        height={450}
+                        quality={90}
+                      />
+                    ),
+                  })}
+                />
+              </Box>
+            );
+          })}
+      </Grid>
+      {!items && children && <>{children}</>}
     </Box>
   );
 }