+++ CHANGELOG.md
@@ -2,7 +2,20 @@
 
 ---
 
-### v0.9.2 [?] ?
+### v0.9.3 - Section Front Research Release
+
+---
+
+- Adds External Search component: [#358](https://github.com/NYPL/dxp-react-search/pull/358)
+- Adds Email Subscription component: [#358](https://github.com/NYPL/dxp-react-search/pull/358)
+- Adds Salesforce API route: [#358](https://github.com/NYPL/dxp-react-search/pull/358)
+- Updates AWS deploy script with Salesforce env variables: [#384](https://github.com/NYPL/dxp-react-search/pull/384)
+- Updates Card Grid component with additional layout variants: [#371](https://github.com/NYPL/dxp-react-search/pull/371)
+- Adds get-colorway util for dynamically setting section front colors: [#353](https://github.com/NYPL/dxp-react-search/pull/353)
+
+---
+
+### v0.9.2 - General Maintenance & Improvements
 
 ---
 
@@ -11,6 +24,8 @@
 - Update Scout to Use DS AlphabetFilter Component and DS 1.2: [#345](https://github.com/NYPL/dxp-react-search/pull/345)
 - Setup Additional AWS environments: [#356](https://github.com/NYPL/dxp-react-search/pull/356)
 - Add new maintenace page and improved error handling: [#343](https://github.com/NYPL/dxp-react-search/pull/343)
+- Revert Global Header for Vega Change: [#368](https://github.com/NYPL/dxp-react-search/pull/368)
+- Update Global Header for Vega Change Take 2: [#370](https://github.com/NYPL/dxp-react-search/pull/370)
 
 ---